docker-rm/requirements.txt
connexion>=1.0.31
gevent>=1.1b1
docker>=2.4.2
kafka-python>=1.3.4
pytz>=2017.2
tinydb>=3.4.1
